+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## [1.3.1-dev.4](https://github.com/RaymondSalim/morphe-patches/compare/v1.3.1-dev.3...v1.3.1-dev.4) (2026-09-14)
+
+### 🐛 Bug Fixes
+
+* Tune guide simulation step from 0.5s to 0.375s ([a7c5d47](https://github.com/RaymondSalim/morphe-patches/commit/a7c5d470d3eabfe2fd59f6295a1d0e3f0f6ec3bc))
+
+## [1.3.1-dev.3](https://github.com/RaymondSalim/morphe-patches/compare/v1.3.1-dev.2...v1.3.1-dev.3) (2026-09-14)
+
+### 🐛 Bug Fixes
+
+* Force guide simulation step to 0.5s so the aim guide reaches the ground ([901686a](https://github.com/RaymondSalim/morphe-patches/commit/901686a70c9a385f59f1bee3be661fcc9fbe09e9))
+
+## [1.3.1-dev.2](https://github.com/RaymondSalim/morphe-patches/compare/v1.3.1-dev.1...v1.3.1-dev.2) (2026-09-14)
+
+### 🐛 Bug Fixes
+
+* Add temporary guide renderer diagnostic patch ([121d80d](https://github.com/RaymondSalim/morphe-patches/commit/121d80dbdd4ac713cdbd323a9fa554af975c1b89))
+* Add temporary guide renderer diagnostic patch ([13be484](https://github.com/RaymondSalim/morphe-patches/commit/13be4844609fbb838596c45e78b22a27da614b59))
+* Add temporary guide renderer diagnostic patch ([e11028e](https://github.com/RaymondSalim/morphe-patches/commit/e11028e40dabdf5cc6583912fc1cf679a1382039))
+* Add temporary guide renderer diagnostic patch ([4321c8f](https://github.com/RaymondSalim/morphe-patches/commit/4321c8f06901918cc5ebdffbb2427fe7e2fa8e0e))
+
+## [1.3.1-dev.1](https://github.com/RaymondSalim/morphe-patches/compare/v1.3.0...v1.3.1-dev.1) (2026-09-14)
+
+### 🐛 Bug Fixes
+
+* Extend in-battle aim guide via SetTrajectoryDots timeStep patch ([7f9f8d8](https://github.com/RaymondSalim/morphe-patches/commit/7f9f8d84bd47d6573e2abb2076a2d9a60e07de69))
+
 ## [1.3.0](https://github.com/RaymondSalim/morphe-patches/compare/v1.2.0...v1.3.0) (2026-09-13)
 
 ### 🐛 Bug Fixes